refactor: auth type 위치 이동
